refactor(spaces): split spaces.rs by responsibility
src/spaces.rs — the core AST->FuncSpace plumbing every metric flows
through — held the public types, the metric-computation/traversal free
functions, and a large in-file test module at 2112 production SLOC,
2.6x the 800-line self-scan limit. Split it by responsibility:

  - spaces/compute.rs  the computation/traversal free functions
                       (analyze, metrics_inner, compute_*, finalize*,
                       push_*, apply_suppression)
  - spaces/ast.rs      impl Ast (+ Debug)
  - spaces/code_metrics.rs  the CodeMetrics impls + Display
  - spaces/options.rs / source.rs / space_kind.rs  the remaining
                       per-type inherent + trait impls
  - spaces_tests.rs    the inline test module, externalized via
                       #[cfg(test)] #[path] (the lib_tests.rs /
                       commands_tests.rs precedent; .bcaignore excludes
                       *_tests.rs)

Pure code movement: bodies verbatim, no logic or metric-output change.
The public type DEFINITIONS (SpaceKind, CodeMetrics, FuncSpace, Source,
Ast, MetricsOptions, State) stay in spaces.rs; their impl blocks live
in child modules (method resolution is by type, not module, so
`spaces::Ast::parse` etc. resolve unchanged). `analyze` (pub) and
`metrics_inner` (pub(crate)) are re-exported from spaces.rs so lib.rs's
`pub use crate::spaces::{...analyze}` / `pub(crate) use ...metrics_inner`
stay byte-identical. No struct field visibility was widened. lib.rs is
unchanged.

spaces.rs 2112 -> 469 production SLOC (genuinely under 800, baseline
entry retired); every child module under 800; tests run verbatim from
spaces_tests.rs.
